requirements.txt
Pillow
pystray
keyboard
